feat: format API definition for readability
When using `$text` placeholders for API entities targeting autogenerated
OpenAPI specs like the ones generated by [spring-boot-open-api](https://springdoc.org/)
default behaviour the output is not very readable (and breaks the layout).

This will check if the definition is a one-liner, will assume it is not pretty printed
then formats it with a notice.